[Step2] refactor: 변수와 메서드명 수정, 패키지 이동, 메인함수에서 로직이 들어가지 않도록 분리
