+++ src/main/kotlin/lotto/agency/LottoTicket.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.agency
 
 class LottoTicket(val numbers: List<Int>)
