+++ src/main/kotlin/lotto/view/InputView.kt
@@ -8,7 +8,7 @@ class InputView {
         return readln()
     }
 
-    fun enterWonLottoLastWeek(): List<String> {
+    fun enterWonLotto(): List<String> {
         println("지난 주 당첨 번호를 입력해 주세요.")
 
         return readln().split(", ")
